코틀린 학습 9월 22일 권현욱

+++ KotlinStudy/Hyun/Programmers Problems/정수 제곱근 판별.ws.kts
@@ -11,7 +11,7 @@ var nSqrt : Double = sqrt(n.toDouble())
 if (nSqrt % 1.0 == 0.0) {
     nSqrt += 1.0
     nSqrt = nSqrt.pow(2)
-    println(nSqrt)
+    println(nSqrt.toInt())
 }
 else {
     println(-1)
